+++ src/bench/descriptors.cpp
@@ -6,7 +6,6 @@
 #include <key.h>
 #include <pubkey.h>
 #include <script/descriptor.h>
-#include <script/standard.h>
 
 #include <string>
 #include <utility>
