+++ src/key_io.h
@@ -6,6 +6,7 @@
 #ifndef BITCOIN_KEY_IO_H
 #define BITCOIN_KEY_IO_H
 
+#include <addresstype.h>
 #include <chainparams.h>
 #include <key.h>
 #include <pubkey.h>
