+++ src/wallet/test/util.h
@@ -5,7 +5,7 @@
 #ifndef BITCOIN_WALLET_TEST_UTIL_H
 #define BITCOIN_WALLET_TEST_UTIL_H
 
-#include <script/standard.h>
+#include <addresstype.h>
 #include <wallet/db.h>
 
 #include <memory>
